refactor(채팅): 디렉토리 구조, 폴더 이름 변경
