+++ backend/chat/HELP.md
