+++ backend/chat/settings.gradle
